Move the encounter API to the programs package
