+++ client/packages/programs/src/api/api.ts
@@ -5,16 +5,20 @@ import {
   DocumentRegistryNodeContext,
   DocumentRegistrySortFieldInput,
   EncounterSortFieldInput,
+  InsertEncounterInput,
   InsertProgramEnrolmentInput,
   ProgramEnrolmentSortFieldInput,
+  UpdateEncounterInput,
   UpdateProgramEnrolmentInput,
 } from '@common/types';
+import { EncounterListParams } from './hooks/utils/useEncounterApi';
 import {
   DocumentFragment,
   DocumentRegistryFragment,
   DocumentRegistryWithChildrenFragment,
-  EncounterBaseFragment,
   EncounterFieldsFragment,
+  EncounterFragment,
+  EncounterRowFragment,
   ProgramEnrolmentRowFragment,
   Sdk,
 } from './operations.generated';
@@ -49,6 +53,15 @@ export const getDocumentQueries = (sdk: Sdk, storeId: string) => ({
       }
       throw new Error('Patient document does not exist');
     },
+    documentHistory: async (
+      documentName: string
+    ): Promise<DocumentFragment[]> => {
+      const result = await sdk.getDocumentHistory({
+        storeId,
+        name: documentName,
+      });
+      return result.documentHistory.nodes;
+    },
   },
 });
 
@@ -58,6 +71,25 @@ export type DocumentRegistryParams = {
 };
 
 export const getEncounterQueries = (sdk: Sdk, storeId: string) => ({
+  list: async ({
+    sortBy,
+    filterBy,
+    pagination,
+  }: EncounterListParams): Promise<{
+    nodes: EncounterRowFragment[];
+    totalCount: number;
+  }> => {
+    const result = await sdk.encounters({
+      storeId,
+      key: sortBy?.key as EncounterSortFieldInput | undefined,
+      desc: sortBy?.isDesc,
+      filter: filterBy,
+      page: pagination,
+      eventTime: new Date().toISOString(),
+    });
+
+    return result?.encounters;
+  },
   encounterFields: async (
     patientId: string,
     fields: string[]
@@ -70,7 +102,7 @@ export const getEncounterQueries = (sdk: Sdk, storeId: string) => ({
     }
     throw new Error('Error querying document');
   },
-  byId: async (encounterId: string): Promise<EncounterBaseFragment> => {
+  byId: async (encounterId: string): Promise<EncounterFragment> => {
     const result = await sdk.encounterById({ encounterId, storeId });
     const encounters = result?.encounters;
 
@@ -86,8 +118,8 @@ export const getEncounterQueries = (sdk: Sdk, storeId: string) => ({
   previousEncounters: async (
     patientId: string,
     current: Date
-  ): Promise<EncounterBaseFragment> => {
-    const result = await sdk.encounters({
+  ): Promise<EncounterFragment> => {
+    const result = await sdk.encountersWithDocument({
       storeId,
       key: EncounterSortFieldInput.StartDatetime,
       desc: true,
@@ -111,6 +143,36 @@ export const getEncounterQueries = (sdk: Sdk, storeId: string) => ({
     } else {
       throw new Error('Could not find encounter');
     }
+  },
+
+  insertEncounter: async (
+    input: InsertEncounterInput
+  ): Promise<EncounterFragment> => {
+    const result = await sdk.insertEncounter({
+      storeId,
+      input,
+    });
+
+    if (result.insertEncounter.__typename === 'EncounterNode') {
+      return result.insertEncounter;
+    }
+
+    throw new Error('Could not insert encounter');
+  },
+
+  updateEncounter: async (
+    input: UpdateEncounterInput
+  ): Promise<EncounterFragment> => {
+    const result = await sdk.updateEncounter({
+      storeId,
+      input,
+    });
+
+    if (result.updateEncounter.__typename === 'EncounterNode') {
+      return result.updateEncounter;
+    }
+
+    throw new Error('Could not update encounter');
   },
 });
