+++ client/packages/programs/src/api/hooks/index.ts
@@ -1,9 +1,21 @@
+import {
+  EncounterFragment,
+  EncounterRowFragment,
+} from '../operations.generated';
 import { Document } from './document';
 import { ProgramEnrolmentRowFragmentWithId } from './document/useProgramEnrolments';
+import { EncounterRegistryByProgram } from './document/useEncounterRegistriesByPrograms';
 import { useDocumentRegistryApi } from './utils/useDocumentRegistryApi';
 import { useEncounterIdFromUrl } from './utils/useEncounterIdFromUrl';
 
-export { ProgramEnrolmentRowFragmentWithId, useDocumentRegistryApi };
+export { EncounterRegistryByProgram };
+
+export {
+  ProgramEnrolmentRowFragmentWithId,
+  useDocumentRegistryApi,
+  EncounterFragment,
+  EncounterRowFragment,
+};
 
 export const useDocument = {
   utils: {
@@ -11,6 +23,7 @@ export const useDocument = {
   },
   get: {
     documentByName: Document.useDocumentByName,
+    history: Document.useDocumentHistory,
   },
 };
 
@@ -18,6 +31,7 @@ export const useDocumentRegistry = {
   get: {
     documentRegistry: Document.useDocumentRegistryByContext,
     programRegistries: Document.useProgramRegistries,
+    encounterRegistriesByPrograms: Document.useEncounterRegistriesByPrograms,
   },
 };
 
@@ -33,8 +47,11 @@ export const useEncounter = {
   },
   encounterFields: Document.useEncounterFields,
   document: {
+    list: Document.useEncounters,
     byId: Document.useEncounterById,
+    byIdPromise: Document.useEncounterByIdPromise,
     previous: Document.useEncounterPrevious,
+    upsert: Document.useUpsertEncounter,
   },
 };
