+++ client/packages/programs/src/api/hooks/utils/useDocumentApi.ts
@@ -6,6 +6,8 @@ export const usePatientDocumentApi = () => {
   const { storeId } = useAuthContext();
   const keys = {
     base: () => ['patient', storeId] as const,
+    history: (docName: string) =>
+      [...keys.base(), 'doc-history', docName] as const,
     byPatient: (patientId: string) => [...keys.base(), 'p', patientId] as const,
     byName: (name: string) => [...keys.base(), 'n', name] as const,
   };
