+++ client/packages/programs/src/api/hooks/utils/useEncounterApi.ts
@@ -6,12 +6,12 @@ import {
   SortBy,
 } from '@openmsupply-client/common';
 import { getEncounterQueries } from '../../api';
-import { EncounterBaseFragment, getSdk } from '../../operations.generated';
+import { EncounterFragment, getSdk } from '../../operations.generated';
 
-export type ListParams = {
+export type EncounterListParams = {
   first?: number;
   offset?: number;
-  sortBy?: SortBy<EncounterBaseFragment>;
+  sortBy?: SortBy<EncounterFragment>;
   filterBy?: FilterBy | null;
   pagination?: PaginationInput;
 };
@@ -22,7 +22,7 @@ export const useEncounterApi = () => {
     base: () => ['encounter'] as const,
     detail: (id: string) => [...keys.base(), storeId, id] as const,
     list: () => [...keys.base(), storeId, 'list'] as const,
-    paramList: (params: ListParams) =>
+    paramList: (params: EncounterListParams) =>
       [...keys.base(), storeId, ...keys.list(), params] as const,
     encounterFields: (patientId: string, fields: string[]) =>
       [...keys.base(), storeId, patientId, 'fields', ...fields] as const,
