--- client/packages/system/src/Encounter/api/hooks/registry/index.ts
@@ -1,5 +1,0 @@
-import { useRegistryByProgram } from './useRegistryByProgram';
-
-export const Registry = {
-  useRegistryByProgram,
-};
