--- client/packages/system/src/Encounter/api/hooks/registry/useRegistryByProgram.ts
@@ -1,12 +1,0 @@
-import { useMutation } from '@openmsupply-client/common';
-import { useEncounterApi } from '../utils/useEncounterApi';
-
-export const useRegistryByProgram = () => {
-  const api = useEncounterApi();
-
-  return {
-    ...useMutation(async (parentId: string) =>
-      api.get.registries({ filterBy: { parentId: { equalTo: parentId } } })
-    ),
-  };
-};
