--- client/packages/system/src/Encounter/api/hooks/utils/index.ts
@@ -1,7 +1,0 @@
-import { useEncounterApi } from './useEncounterApi';
-import { useEncounterId } from './useEncounterId';
-
-export const Utils = {
-  useEncounterApi: useEncounterApi,
-  useEncounterId: useEncounterId,
-};
