--- client/packages/system/src/Encounter/api/hooks/utils/useEncounterId.ts
@@ -1,6 +1,0 @@
-import { useParams } from '@openmsupply-client/common';
-
-export const useEncounterId = () => {
-  const { id = '' } = useParams();
-  return decodeURIComponent(id);
-};
