--- client/packages/system/src/Encounter/api/index.ts
@@ -1,9 +1,0 @@
-import {
-  EncounterFragment,
-  EncounterRowFragment,
-} from './operations.generated';
-
-type EncounterFragmentWithId = { id: string } & EncounterFragment;
-
-export * from './hooks';
-export { EncounterFragment, EncounterRowFragment, EncounterFragmentWithId };
