+++ client/packages/system/src/Patient/api/hooks/index.ts
@@ -9,13 +9,10 @@ export const usePatient = {
     search: Utils.usePatientSearch,
   },
   document: {
-    encounters: Document.useEncounters,
     get: Document.usePatient,
     list: Document.usePatients,
     listAll: Document.usePatientsAll,
     insert: Document.useInsertPatient,
     update: Document.useUpdatePatient,
-    history: Document.useDocumentHistory,
-    programEncounters: Document.useProgramEncounters,
   },
 };
